VXFM-11201 when a 3NIC gateway ova is deployed using VXFM,it configures 5NICS in the vcenter on the VM deployed
In case "VM Netwok" default port-group is deleted or renamed, then OVF creation fails with no port-group associated with extra networks.

Considering above, we cannot be sure if there will be a default port-group on vCenter. I am planning to change the behavior

Check for VM Network on the environment. If found, then use it
if VM network is not found then map all extra networks with last network (same as original behavior)